homeworkupload
Hi Nicole! So initially wanted to create an app for a bat that cares about their hairstyle and does not want to fly when it's to humid outside. I wanted have an image of a bat pop up based on humidity (low humidity = happy dry bat; high humidity = sad bat with curly hair) - I tried this by loading two images of the internet/myfolder but it did not work (the lines are still in the file you'll see). The rest of the code works fine and I made conditions that react to the data from the API -unfortunately the images did not work out so I decidede to go for a text based version instead.